rename profile link
