+++ frontend/src/metabase/nav/components/AppBar/AppBarLarge.tsx
@@ -9,8 +9,8 @@ import type { DetailViewState } from "metabase-types/store";
 
 import CollectionBreadcrumbs from "../../containers/CollectionBreadcrumbs";
 import QuestionLineage from "../../containers/QuestionLineage";
+import { AppSwitcher } from "../AppSwitcher";
 import NewItemButton from "../NewItemButton";
-import { ProfileLink } from "../ProfileLink/ProfileLink";
 import { SearchBar } from "../search/SearchBar";
 import { SearchButton } from "../search/SearchButton/SearchButton";
 
@@ -30,7 +30,7 @@ export interface AppBarLargeProps {
   isSearchVisible?: boolean;
   isEmbeddingIframe?: boolean;
   isNewButtonVisible?: boolean;
-  isProfileLinkVisible?: boolean;
+  isAppSwitcherVisible?: boolean;
   isCollectionPathVisible?: boolean;
   isQuestionLineageVisible?: boolean;
   onToggleNavbar: () => void;
@@ -48,7 +48,7 @@ const AppBarLarge = ({
   isSearchVisible,
   isEmbeddingIframe,
   isNewButtonVisible,
-  isProfileLinkVisible,
+  isAppSwitcherVisible,
   isCollectionPathVisible,
   isQuestionLineageVisible,
   onToggleNavbar,
@@ -94,7 +94,7 @@ const AppBarLarge = ({
           ) : null}
         </AppBarInfoContainer>
       </Flex>
-      {(isSearchVisible || isNewButtonVisible || isProfileLinkVisible) && (
+      {(isSearchVisible || isNewButtonVisible || isAppSwitcherVisible) && (
         <Flex
           align="center"
           gap="sm"
@@ -106,9 +106,9 @@ const AppBarLarge = ({
             (isEmbeddingIframe ? <SearchBar /> : <SearchButton mr="md" />)}
           {isNewButtonVisible && <NewItemButton collectionId={collectionId} />}
           {!isEmbeddingIframe && <PLUGIN_METABOT.MetabotAppBarButton />}
-          {isProfileLinkVisible && (
+          {isAppSwitcherVisible && (
             <Box c="text-primary" aria-label={t`Settings menu`}>
-              <ProfileLink />
+              <AppSwitcher />
             </Box>
           )}
         </Flex>
