+++ frontend/src/metabase/nav/components/AppSwitcher/index.ts
@@ -1,0 +1,1 @@
+export * from "./AppSwitcher";
