+++ frontend/src/metabase/nav/components/AppSwitcher/useHelpLink.ts
