+++ frontend/src/metabase/nav/components/AppSwitcher/useHelpLink.unit.spec.ts
