--- frontend/src/metabase/nav/components/ProfileLink/index.ts
@@ -1,1 +1,0 @@
-export * from "./ProfileLink";
